test-only: move testing controller to test-only route
