+++ app/controllers/test/GeneratedXMLController.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.test
 
 import controllers.actions._
 import play.api.i18n.{I18nSupport, MessagesApi}
