+++ conf/app.routes
@@ -589,4 +589,3 @@ POST       /reporter/individual/change-select-address        controllers.reporte
 # TASK LIST
 POST      /task-list                                         controllers.TaskListController.onSubmit
 
-GET       /generated-xml                                     controllers.GeneratedXMLController.onPageLoad()
