+++ conf/testOnlyDoNotUseInAppConf.routes
@@ -11,3 +11,5 @@
 
 # Add all the application routes to the prod.routes file
 ->         /                          prod.Routes
+
+GET       /test-only/generated-xml                                     controllers.test.GeneratedXMLController.onPageLoad()
